requirements/base.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
autopep8==2.0.2
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.4
django-rest-framework==0.1.0
django-silk==5.0.3
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
drf-yasg==1.21.7
flower==2.0.1
gprof2dot==2022.7.29
gunicorn==21.2.0
humanize==4.8.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.1
MarkupSafe==2.1.3
packaging==23.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg==3.1.10
pycodestyle==2.11.0
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
requests==2.31.0
sentry-sdk==1.29.2
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tornado==6.3.3
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
